test(pipe): move plugin load failure test out of PowerMock
